refactor: split in files hold amount function
